split `correlator`
